refs #104: atomファイルのtypeをtypeファイルに置き換える
