+++ src/hooks/useDatasets.ts
@@ -1,4 +1,5 @@
-import { dataset, githubUsers } from '@/store/atoms';
+import { dataset } from '@/types/DatasetType';
+import { githubUsers } from '@/types/GitHubApiType';
 
 export const useDatasets = (githubUsers: githubUsers) => {
   const getDatasets = githubUsers.map((user) => {
