+++ src/pages/project.tsx
@@ -9,11 +9,11 @@ import { useAuth } from '@/context/auth';
 import { useGithubUsers } from '@/hooks/useGithubUsers';
 import { getShowUsersHandler } from '@/lib/firebase/getShowUsersHandler';
 import { db } from '@/lib/firebaseClient';
-import { githubUsers } from '@/store/atoms';
 import { fetchProjectState } from '@/store/fetchProjectAtoms';
 import { memberCountState } from '@/store/memberCountAtoms';
 import { projectSelectViewState } from '@/store/projectSelectViewAtoms';
 import { isOpenSidebarState } from '@/store/sidebarAtoms';
+import { githubUsers } from '@/types/GitHubApiType';
 import { NonIdProjectType } from '@/types/ProjectType';
 import { ShowUserType } from '@/types/ShowUserType';
 import { css } from '@emotion/react';
