+++ src/types/DatasetType.ts
@@ -1,0 +1,6 @@
+export type dataset = {
+  label: string;
+  data: (number | undefined)[];
+  borderColor: string;
+  backgroundColor: string;
+}[];
